refactor: extract upload entry validation to sharing-server/validation/uploadSchema.ts
Move validateEntry() and its isNonNegativeInt() helper out of api.ts into a
dedicated validation module. Magic numbers become named constants:
- MAX_INTERACTIONS (100_000) — was an inline literal
- MAX_FLUENCY_METRICS_JSON_BYTES (10_000) — was a module-level constant in api.ts

Enables reuse in tests and future endpoints. api.ts now imports validateEntry
from the new module; fluency-score validation remains in api.ts as it belongs
to a separate endpoint.
